Move generate_funs() out of utilities to resolve the layering inversion
`utilities.generate_funs()` reached up to `compactfun.CompactFun` through a
function-local import — the only upward dependency in the package. The
deferred import was a working workaround for the inversion, not a fix: the
function dispatches between `Bndfun` and `CompactFun`, so it cannot honestly
live in the primitives layer.

Move it to a new `_construction.py` alongside `_singular_construction.py`,
which already does the analogous job for domains with endpoint singularities.
Both now sit at the Chebfun impl-module layer, where knowing about the
Classicfun pieces is permitted, and `CompactFun` imports at module scope.

`utilities.py` no longer imports `Callable`, now unused there. Tests move to
`tests/test_construction.py`, extended to cover the finite, unbounded and
default-domain paths rather than only the rejection guard.

Closes #475
Closes #417
